requirements.txt
openai
python-dotenv
playwright
nest-asyncio
asyncio
jinja2
flask==2.1.3
flask-cors
aioflask
werkzeug==2.2.2
requests
pillow
bs4
